LPD-85827 Extract isContentStructureMoveInvalid to util file
